\name{dfmta-package}
\alias{dfmta-package}
\alias{dfmta}
\docType{package}
\title{
Phase I/II Dose-Finding Design for Molecularly Targeted Agent
}
\description{
This package provides functions to run simulations and to determine the next 
optimal dose in on-going trials for Phase I/II dose-finding design for 
Molecularly Targeted Agent.
}
\details{
\tabular{ll}{
Package: \tab dfmta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-23\cr
License: \tab GPL-3\cr
}
}
\author{
Jacques-Henri Jourdan and Marie-Karelle Riviere-Jourdan <eldamjh@gmail.com>
}
\references{
Riviere, M-K., Yuan, Y., Jourdan, J-H., Dubois, F., and Zohar, S. (2014). Phase I/II Dose-Finding Design for Molecularly Targeted Agent: Plateau Determination using Adaptive Randomization. Statistics in medicine.
}
