\name{dfmta-package}
\alias{dfmta-package}
\alias{dfmta}
\docType{package}
\title{
\packageTitle{dfmta}
}
\description{
\packageDescription{dfmta}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dfmta}
\packageIndices{dfmta}
}
\author{
\packageAuthor{dfmta}

Maintainer: \packageMaintainer{dfmta}
}
\references{
Riviere, M-K., Yuan, Y., Jourdan, J-H., Dubois, F., and Zohar, S. Phase I/II Dose-Finding Design for Molecularly Targeted Agent: Plateau Determination using Adaptive Randomization.
}
\keyword{ package design survival }
